pyproject.toml
[project]
name = "monobit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "django>=5.2.11",
    "djangorestframework>=3.16.1",
    "gunicorn>=25.1.0",
    "psycopg>=3.3.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
]
